lustre: llite: move agl_thread cleanup out of thread.
When we start a thread with kthread_create() and later stop
it with kthread_stop(), there is no guarantee that the thread
function runs at all.  So it is not safe to leave cleanup
to the thread.

So move the cleanup code to a separate function which
stops the thread and then cleans up.

Fixes: c044fb0f835c ("staging: lustre: remove 'ptlrpc_thread usage' for sai_agl_thread")

WC-bug-id: https://jira.whamcloud.com/browse/LU-12780
Lustre-commit: 5b630935452d ("LU-12780 llite: don't use ptlrpc_thread for sai_agl_thread")
Signed-off-by: NeilBrown <neilb@suse.de>
Reviewed-on: https://review.whamcloud.com/36258
Reviewed-by: Lai Siyao <lai.siyao@whamcloud.com>
Reviewed-by: James Simmons <jsimmons@infradead.org>
Reviewed-by: Shaun Tancheff <shaun.tancheff@hpe.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Signed-off-by: James Simmons <jsimmons@infradaed.org>